SLEMA-82 Unify and separate default container
