+++ SlemaForAndroid/lib/features/home/presentation/screen/home_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:pg_slema/features/home/presentation/widget/home_app_bar.dart';
-import 'package:pg_slema/features/home/presentation/widget/home_widget_custom_container.dart';
 import 'package:pg_slema/features/home/presentation/widget/labeled_divider.dart';
 import 'package:pg_slema/features/motivation/presentation/widget/motivation_daily.dart';
+import 'package:pg_slema/utils/widgets/default_container/default_container.dart';
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({super.key});
@@ -25,39 +25,42 @@ class HomeScreenState extends State<HomeScreen> {
         ),
         const LabeledDivider(label: "Dzienny raport zdrowotny"),
         const SizedBox(
-          height: 20.0,
+          height: 5.0,
         ),
-        HomeWidgetCustomContainer(
-            child: Text(
-          "Raport zdrowotny",
-          textAlign: TextAlign.center,
-          style: Theme.of(context).textTheme.labelMedium!.copyWith(
-                color: Theme.of(context).primaryColor,
-              ),
-        )),
+        DefaultContainer(
+          shadow: false,
+          child: Text(
+            "Raport zdrowotny",
+            textAlign: TextAlign.center,
+            style: Theme.of(context).textTheme.labelMedium!.copyWith(
+                  color: Theme.of(context).primaryColor,
+                ),
+          ),
+        ),
         const SizedBox(
-          height: 20.0,
+          height: 5.0,
         ),
         const LabeledDivider(label: "Nadchodzące wydarzenia"),
         const SizedBox(
-          height: 20.0,
+          height: 5.0,
         ),
-        HomeWidgetCustomContainer(
+        DefaultContainer(
+            shadow: false,
             child: Text(
-          "Kalendarz",
-          textAlign: TextAlign.center,
-          style: Theme.of(context).textTheme.labelMedium!.copyWith(
-                color: Theme.of(context).primaryColor,
-              ),
-        )),
+              "Kalendarz",
+              textAlign: TextAlign.center,
+              style: Theme.of(context).textTheme.labelMedium!.copyWith(
+                    color: Theme.of(context).primaryColor,
+                  ),
+            )),
         const SizedBox(
-          height: 20.0,
+          height: 5.0,
         ),
         const LabeledDivider(label: "Dla ciebie"),
         const SizedBox(
-          height: 20.0,
+          height: 5.0,
         ),
-        const HomeWidgetCustomContainer(child: MotivationDaily()),
+        const DefaultContainer(shadow: false, child: MotivationDaily()),
       ],
     );
   }
