+++ SlemaForAndroid/lib/features/medicine/presentation/widget/all_medicines_screen/single_medicine_widget.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:pg_slema/features/medicine/logic/entity/medicine.dart';
 import 'package:pg_slema/features/medicine/presentation/widget/all_medicines_screen/single_medicine_label.dart';
 import 'package:pg_slema/features/medicine/presentation/widget/all_medicines_screen/popup_menu_edit_delete_button.dart';
+import 'package:pg_slema/utils/widgets/default_container/default_container.dart';
 
 class SingleMedicineWidget extends StatelessWidget {
   final ValueChanged<Medicine> onMedicineDeleted;
@@ -16,70 +17,54 @@ class SingleMedicineWidget extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-        margin: const EdgeInsets.symmetric(horizontal: 15.0, vertical: 15.0),
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(12.0),
-          color: Theme.of(context).colorScheme.primaryContainer,
-          boxShadow: [
-            BoxShadow(
-              color: Theme.of(context).colorScheme.shadow,
-              offset: const Offset(0.0, 4.0),
-              blurRadius: 4.0,
+    return DefaultContainer(
+      child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
+        Row(
+          mainAxisAlignment: MainAxisAlignment.end,
+          children: [
+            Padding(
+              padding: const EdgeInsets.only(bottom: 8.0, right: 8.0),
+              child: Icon(
+                Icons.medication_outlined,
+                color: Theme.of(context).primaryColor,
+                size: 32,
+              ),
             ),
+            Expanded(
+              child: Text(
+                medicine.name,
+                style: Theme.of(context).textTheme.headlineLarge,
+                overflow: TextOverflow.ellipsis,
+              ),
+            ),
+            PopupMenuEditDeleteButton(
+                onMedicineChanged: onMedicineEdited,
+                medicineProvider: () => medicine,
+                onDeletePressed: () => onMedicineDeleted(medicine)),
           ],
-          border: Border.all(
-            color: Theme.of(context).colorScheme.outline,
-            width: 3.0,
-          ),
         ),
-        padding: const EdgeInsets.symmetric(vertical: 5.0, horizontal: 10.0),
-        child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.end,
-            children: [
-              Padding(
-                padding: const EdgeInsets.only(bottom: 8.0, right: 8.0),
-                child: Icon(
-                  Icons.medication_outlined,
-                  color: Theme.of(context).primaryColor,
-                  size: 32,
-                ),
+        Wrap(
+          runSpacing: 5.0,
+          spacing: 5.0,
+          children: [
+            if (medicine.dose.isNotEmpty)
+              SingleMedicineLabel(
+                label: medicine.dose,
+                color: const Color(0xFFB6DDFA),
               ),
-              Expanded(
-                child: Text(
-                  medicine.name,
-                  style: Theme.of(context).textTheme.headlineLarge,
-                  overflow: TextOverflow.ellipsis,
-                ),
+            if (medicine.medicineType.isNotEmpty)
+              SingleMedicineLabel(
+                label: medicine.medicineType,
+                color: const Color(0xFFFAB6B6),
               ),
-              PopupMenuEditDeleteButton(
-                  onMedicineChanged: onMedicineEdited,
-                  medicineProvider: () => medicine,
-                  onDeletePressed: () => onMedicineDeleted(medicine)),
-            ],
-          ),
-          Wrap(
-            runSpacing: 5.0,
-            spacing: 5.0,
-            children: [
-              if (medicine.dose.isNotEmpty)
-                SingleMedicineLabel(
-                  label: medicine.dose,
-                  color: const Color(0xFFB6DDFA),
-                ),
-              if (medicine.medicineType.isNotEmpty)
-                SingleMedicineLabel(
-                  label: medicine.medicineType,
-                  color: const Color(0xFFFAB6B6),
-                ),
-              if (medicine.intakeType.isNotEmpty)
-                SingleMedicineLabel(
-                  label: medicine.intakeType,
-                  color: const Color(0xFFB6FAB6),
-                ),
-            ],
-          )
-        ]));
+            if (medicine.intakeType.isNotEmpty)
+              SingleMedicineLabel(
+                label: medicine.intakeType,
+                color: const Color(0xFFB6FAB6),
+              ),
+          ],
+        )
+      ]),
+    );
   }
 }
