+++ SlemaForAndroid/lib/features/well_being_assessment/presentation/screen/assessment_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:pg_slema/features/well_being_assessment/presentation/widget/assessment_app_bar.dart';
 import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/sleep/assessment_sleep_form.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/common/assessment_entry.dart';
+import 'package:pg_slema/utils/widgets/default_container/default_container.dart';
 import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/symptoms/assessment_symptoms_form.dart';
 import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/well_being/assessment_well_being_form.dart';
 import 'package:pg_slema/utils/date/date.dart';
@@ -28,13 +28,13 @@ class _AssessmentScreenState extends State<AssessmentScreen> with Logger {
             key: _formKey,
             child: const Column(
               children: [
-                AssessmentFormContainer(
+                DefaultContainer(
                   child: AssessmentWellBeingForm(),
                 ),
-                AssessmentFormContainer(
+                DefaultContainer(
                   child: AssessmentSymptomsFormWidget(),
                 ),
-                AssessmentFormContainer(
+                DefaultContainer(
                   child: AssessmentSleepForm(),
                 ),
               ],
